requirements.txt
boto3==1.26.147
botocore==1.29.147
celery==5.3.0
Django==4.2.2
django-storages==1.13.2
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.24.3
Pillow==9.5.0
PyMuPDF==1.22.3
python-dotenv==1.0.0
requests==2.31.0